+++ PoshBot/Classes/Approver.ps1
@@ -1,0 +1,5 @@
+
+class Approver {
+    [string]$Id
+    [string]$Name
+}