+++ PoshBot/Classes/CommandAuthorizationResult.ps1
@@ -1,0 +1,18 @@
+
+class CommandAuthorizationResult {
+    [bool]$Authorized
+    [string]$Message
+
+    CommandAuthorizationResult() {
+        $this.Authorized = $true
+    }
+
+    CommandAuthorizationResult([bool]$Authorized) {
+        $this.Authorized = $Authorized
+    }
+
+    CommandAuthorizationResult([bool]$Authorized, [string]$Message) {
+        $this.Authorized = $Authorized
+        $this.Message = $Message
+    }
+}