+++ PoshBot/Classes/CommandResult.ps1
@@ -1,12 +1,4 @@
 
-class Stream {
-    [object[]]$Debug = @()
-    [object[]]$Error = @()
-    [object[]]$Information = @()
-    [object[]]$Verbose = @()
-    [object[]]$Warning = @()
-}
-
 # Represents the result of running a command
 class CommandResult {
     [bool]$Success