+++ PoshBot/Classes/Stream.ps1
@@ -1,0 +1,8 @@
+
+class Stream {
+    [object[]]$Debug = @()
+    [object[]]$Error = @()
+    [object[]]$Information = @()
+    [object[]]$Verbose = @()
+    [object[]]$Warning = @()
+}