+++ PoshBot/Public/New-PoshBotConfiguration.ps1
@@ -97,6 +97,23 @@ function New-PoshBotConfiguration {
         Instead of PoshBot returning a warning message when it is unable to find a command, use this to parameter to tell PoshBot to return nothing.
     .PARAMETER AddCommandReactions
         Add reactions to a chat message indicating the command is being executed, has succeeded, or failed.
+    .PARAMETER ApprovalExpireMinutes
+        The amount of time (minutes) that a command the requires approval will be pending until it expires.
+    .PARAMETER ApprovalCommandConfigurations
+        Array of hashtables containing command approval configurations.
+
+        @(
+            @{
+                Expression = 'MyModule:Execute-Deploy:*'
+                Groups = 'platform-admins'
+                PeerApproval = $true
+            }
+            @{
+                Expression = 'MyModule:Deploy-HRApp:*'
+                Groups = @('platform-managers', 'hr-managers')
+                PeerApproval = $true
+            }
+        )
     .EXAMPLE
         PS C:\> New-PoshBotConfiguration -Name Cherry2000 -AlternateCommandPrefixes @('Cherry', 'Sam')
 
@@ -116,6 +133,7 @@ function New-PoshBotConfiguration {
         SendCommandResponseToPrivate     : {}
         MuteUnknownCommand               : False
         AddCommandReactions              : True
+        ApprovalConfiguration            : ApprovalConfiguration
 
         Create a new PoshBot configuration with default values except for the bot name and alternate command prefixes that it will listen for.
     .EXAMPLE
@@ -145,6 +163,7 @@ function New-PoshBotConfiguration {
         SendCommandResponseToPrivate     : {}
         MuteUnknownCommand               : False
         AddCommandReactions              : True
+        ApprovalConfiguration            : ApprovalConfiguration
 
         PS C:\> $myBotConfig | Start-PoshBot -AsJob
 
@@ -184,7 +203,10 @@ function New-PoshBotConfiguration {
         [char[]]$AlternateCommandPrefixSeperators = @(':', ',', ';'),
         [string[]]$SendCommandResponseToPrivate = @(),
         [bool]$MuteUnknownCommand = $false,
-        [bool]$AddCommandReactions = $true
+        [bool]$AddCommandReactions = $true,
+        [int]$ApprovalExpireMinutes = 30,
+        [hashtable[]]$ApprovalCommandConfigurations = @()
+
     )
 
     Write-Verbose -Message 'Creating new PoshBot configuration'
@@ -210,6 +232,16 @@ function New-PoshBotConfiguration {
     $config.PluginRepository = $PluginRepository
     $config.SendCommandResponseToPrivate = $SendCommandResponseToPrivate
     $config.AddCommandReactions = $AddCommandReactions
+    $config.ApprovalConfiguration.ExpireMinutes = $ApprovalExpireMinutes
+    if ($ApprovalCommandConfigurations.Count -ge 1) {
+        foreach ($item in $ApprovalCommandConfigurations) {
+            $acc = [ApprovalCommandConfiguration]::new()
+            $acc.PluginCommandExpression = $item.Expression
+            $acc.ApprovalGroups = $item.Groups
+            $acc.PeerApproval = $item.PeerApproval
+            $config.ApprovalConfiguration.Commands.Add($acc) > $null
+        }
+    }
 
     $config
 }